backend/pyproject.toml
[project]
name = "shooting-range-backend"
version = "1.0.0"
description = "Multi-Lane Shooting Range Scoring System Backend"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2,<5.0",
    "djangorestframework>=3.14,<4.0",
    "channels>=4.0,<5.0",
    "channels-redis>=4.1,<5.0",
    "daphne>=4.0,<5.0",
    "gunicorn>=21.0,<22.0",
    "uvicorn[standard]>=0.25,<0.26",
    "redis>=5.0,<6.0",
    "psycopg2-binary>=2.9,<3.0",
    "python-dotenv>=1.0,<2.0",
    "pyjwt>=2.8,<3.0",
    "cryptography>=41.0,<42.0",
    "pydantic>=2.5,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<8.0",
    "pytest-django>=4.7,<5.0",
    "pytest-asyncio>=0.23,<0.24",
    "black>=23.12,<24.0",
    "ruff>=0.1,<0.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["shooting_range*"]

[tool.ruff]
line-length = 100
target-version = "py311"
